fix: ignore empty selector to avoid restarting all pods
set -e is being used to force the script to stop when an error occur.
This seems to not be working when running commands inside a subshell,
which causes the restart of all pods in that namespace.

This is a serious bug and if I was not quick enough or ran the command
in an unsuppervised manner it would restart all our services in production :(.